#108: renaming and removing files
