+++ PYTHON_CDM_Conversion_code/hourly_qff_to_cdm_core.py
