--- PYTHON_CDM_Conversion_code/psv_to_parquet_pipeline_summary.pdf
@@ -1,86 +1,0 @@
-%PDF-1.4
-%“Œ‹ž ReportLab Generated PDF document http://www.reportlab.com
-1 0 obj
-<<
-/F1 2 0 R /F2 3 0 R /F3 4 0 R /F4 5 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/BaseFont /Helvetica-Bold /Encoding /WinAnsiEncoding /Name /F2 /Subtype /Type1 /Type /Font
->>
-endobj
-4 0 obj
-<<
-/BaseFont /Symbol /Name /F3 /Subtype /Type1 /Type /Font
->>
-endobj
-5 0 obj
-<<
-/BaseFont /Times-Roman /Encoding /WinAnsiEncoding /Name /F4 /Subtype /Type1 /Type /Font
->>
-endobj
-6 0 obj
-<<
-/Contents 10 0 R /MediaBox [ 0 0 595.2756 841.8898 ] /Parent 9 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->> 
-  /Type /Page
->>
-endobj
-7 0 obj
-<<
-/PageMode /UseNone /Pages 9 0 R /Type /Catalog
->>
-endobj
-8 0 obj
-<<
-/Author (\(anonymous\)) /CreationDate (D:20250827084034+00'00') /Creator (\(unspecified\)) /Keywords () /ModDate (D:20250827084034+00'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
-  /Subject (\(unspecified\)) /Title (\(anonymous\)) /Trapped /False
->>
-endobj
-9 0 obj
-<<
-/Count 1 /Kids [ 6 0 R ] /Type /Pages
->>
-endobj
-10 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 1676
->>
-stream
-Gau0C>ur41&:O:Se0E7*9/fSon2BR9a'!%dPhD,(m,>l05d\9s!f[$B[jl+'5oJ$4lf"e(UiLD0\lF_B2"B[rhbaGo5(p8V$a'`P.Cd+P_seR.%8R`eDK2rq6H=;@0^ON2JK%=i+_N/>l^$.2@"MI*i.c?=kKu^ai4.Sc`9lLdC4j.Ui"UnVGc\;oX>]Y'6*[bC5n+#l/s]b[5\:U08u+6[Ln)L0jX+I3!?+$R8H(-RFqa@Ma"J'M[M<nbXlHtf7]+O/)a*A1QGQoQ7BU&u9:4.4NcVTHlYN(E`iHnsT15^tReqmEDX/]si)t<+;U`EcAVP9"5[!S)935A5"XNV;`*@jaF]OB4Qe"LL6"qblNIc/B0];1PP7^T8Wp=WMF>QL=lbLrjL35*,Ee]J^>\Q='Q4e,YR:L<]j+_#A%M.5UnF&JCC4U\=/B$3M&'8XBLQu#11(dH`"2F=Qk]<Wc;\'cmcJ8jt]dg36MjI#8!-?"7,uEi*K7:t*bc!uk&aa;&1!JZ-eU_(_$.E.;KTAkKBYZVS=+]P1RsIoDa.#!0_+\Ia]F^%`3DV"\duBr7.C@Ku5tdIW4l@l[P9;.a*T`$c^c\oF$:8;Y;F=0aGetb4nc^o5=;mc0cH/5<>8^SMV9YF=W+H`bQ6OG8`OhCX[Ud%rAi#VdL,WSUij#IPh,FA?A[YX:^lTC0U!s^kR,jndM]V]43Fpcc\*GmFG>^N$[bfS,k=j.$5.Zs[Mgi0SI-fQs=?l-kN_W*[U3Ft$F,g`>HHXs+TBi0@=D00CX>B[.(7P;^A>W%S:s54X;q![ukN,Ck9S^JiRB/6CPI?99FWIuhI\D1t9G/[@;%OKnXLdKR('f:YFEI;hl59.$[ZKgj*S6U(_Jo]U98F:#5-k5C&JB@=btNc0fKEmQaVRChK+e4R]#f\Q`H0sIUrCI^@ac_KH+FVD9;su^?V=-3":&,OKZ$bM=Ot.Gk%MJ&+7Yk^)&pu)b>$0Fb#?cSG\+'T\-6TA&r\M`:/J:X0e_^MZ*`6B,(Th4)'VHc8f^bdIPpG-F)5Gg]st7PPi_.;BDA%5$2/RtmoN_3l\nW,e1FbF=<E-4FhORE\&,i7d"&hBl]Pu5K+",%7G<r<:CjO(s6A3sfAi?=_nih@q"!^97b;9(b`2:8%u*3g1Pn`?i+jdAXZFE3^C7QFdA7H#Z[3/qN\s3i.^cYDqV"n^jE<QgJI5Fa33&Yk`ddnXce,OX@m&c7ms9(jU[1Oj7=$`<qr\4\eNCbpFukZJPERYCcg3b$q:98,Dnal<p<HiqSk=]NbiZb4F[F.'n!mGULSeg3s4()<mn[nTF*8l))D22_iUAfpde0&5E"rM(N@`FpJ&_.=Ie\jd%jul*NGF^uAnQOAa*n<$J+Wu^iku/ZFq=k.haG.Y]*1s='DNT7DE\2qdabL]5>BW0M/YHS@$5jET7-9VpoE3tL!;R25'=u-K#kL\k_,MRDrKThE^.k&01^Hi1MO>3`S*MM519WUeM[O)R9),":=)E*.(J`W*@sJGq.LCIpRROe26/!9jP[Cr$Su@#kjN_b?HMumb_^K3:?m$SD`aTT&cS;Acu%cl1[#\rlYc3<T9J'HHhb8.hN51l$<6mM1SC?p5GBV"+nA:--h(CMGHZhoJk_d*e#VNa_L78hf6fFs`e5;@E"ceiE:PZ,$ksi~>endstream
-endobj
-xref
-0 11
-0000000000 65535 f 
-0000000073 00000 n 
-0000000134 00000 n 
-0000000241 00000 n 
-0000000353 00000 n 
-0000000430 00000 n 
-0000000539 00000 n 
-0000000743 00000 n 
-0000000811 00000 n 
-0000001094 00000 n 
-0000001153 00000 n 
-trailer
-<<
-/ID 
-[<a1a8ae844f93404d65603e0696b3f856><a1a8ae844f93404d65603e0696b3f856>]
-% ReportLab generated PDF document -- digest (http://www.reportlab.com)
-
-/Info 8 0 R
-/Root 7 0 R
-/Size 11
->>
-startxref
-2921
-%%EOF
